requirements.txt
Django==1.11
django-debug-toolbar==2.1
mistune==0.8.4
mysqlclient==1.4.6
pytz==2019.3
sqlparse==0.3.0
